pyproject.toml
[project]
name = "airfold-apps-e2e"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.4",
    "neon-api>=0.3.0",
    "psycopg2-binary>=2.9.11",
    "pygithub>=2.8.1",
    "pyyaml>=6.0.3",
]
